Split transcript surface parity gates
